+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Version 1.5.1 (not released yet)
 - load V3M meshes from `user_maps\meshes` and `red\meshes` in the level editor
 - fix lights randomly stopping working after copying in the level editor
 - fix rendering of VBM textures from `user_maps\textures` in the level editor
+- fix level editor crash when selecting decal texture from 'All' folder
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------